Add functionality to disable logging and extend GROUPME_ENABLE feature.
Add flag ZULIP_LOG_ENABLE (default True) to enable sending outside of the
PUBLIC_TWO_WAY instances.  This flag needs to be added to the configs of
deployments that require it.

Add more (debug) logging about why we are not sending to groupme.  Use the
GROUPME_ENABLE flag to determine if send_to_groupme is even called.